requirements.txt
fastapi
pillow
aiohttp
